pyproject.toml
[tool.poetry]
name = "studious-fortnight"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.1.0"
pytest = "^8.3.3"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.10"
dataclasses-json = "^0.6.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
